refactor(core): share WebKit-epoch and JSON-file resolution helpers
Address parallel-review simplification findings without changing any
Analyzer output.

- Extract the byte-identical WebKit 1601 epoch helpers (offset constant +
  utcFromUnixMicros) into core/src/webkit-epoch.ts and delete the four
  verbatim copies in bookmarks, cookies, history, and login-data.
- Move the Manifest JSON-file resolver (resolveJsonFile + FileResolution)
  into preferences-json.ts, which already owns the verified-JSON reader,
  and delete the duplicate in bookmarks. This also unifies the
  entry-missing reason vocabulary (bookmarks now emits the same
  *_manifest_entry_missing / *_absent taxonomy as preferences).

Pure refactor: successful analysis results are byte-for-byte identical
(all four epoch copies and both resolver copies were already identical
except for bookmarks' entry-missing reason string, which no test asserts
and which is now consistent with the sibling artifacts). pnpm verify
green (112 tests).
