+++ core/src/cookies.ts
@@ -1,5 +1,9 @@
 import { join } from "node:path";
 
+import {
+  WINDOWS_EPOCH_OFFSET_MICROS,
+  utcFromUnixMicros,
+} from "./forensic-time.js";
 import type {
   CookieArtifactWrite,
   DeclaredOriginOs,
@@ -35,7 +39,6 @@ import { schemeOf, type DecryptionSettings } from "./oscrypt.js";
  * Below it the analyzer refuses to assert an epoch without a Declared Origin OS.
  */
 const COOKIE_VERIFIED_SCHEMA_MINIMUM = 10;
-const WINDOWS_EPOCH_OFFSET_MICROS = 11_644_473_600_000_000n;
 
 const SAME_SITE_LABELS = new Map<bigint, string>([
   [-1n, "unspecified"],
@@ -136,28 +139,6 @@ function enumFields(
         ? unavailableField("unsupported_value")
         : valueField(label),
   };
-}
-
-function floorDivision(value: bigint, divisor: bigint): bigint {
-  const quotient = value / divisor;
-  const remainder = value % divisor;
-  return remainder < 0n ? quotient - 1n : quotient;
-}
-
-function utcFromUnixMicros(unixMicros: bigint): string | null {
-  const seconds = floorDivision(unixMicros, 1_000_000n);
-  const micros = unixMicros - seconds * 1_000_000n;
-  const milliseconds = seconds * 1000n + micros / 1000n;
-  const numericMilliseconds = Number(milliseconds);
-  if (!Number.isSafeInteger(numericMilliseconds)) {
-    return null;
-  }
-  const date = new Date(numericMilliseconds);
-  if (Number.isNaN(date.getTime())) {
-    return null;
-  }
-  const base = date.toISOString();
-  return `${base.slice(0, -5)}.${micros.toString().padStart(6, "0")}Z`;
 }
 
 function resolveCookieEpoch(
