+++ core/src/preferences-json.ts
@@ -1,5 +1,7 @@
 import { lstat } from "node:fs/promises";
+import { join } from "node:path";
 
+import type { CaseSourceRecord } from "./case.js";
 import { WorkingCopyIntegrityRefusal } from "./errors.js";
 import { readStableRegularFile } from "./stable-file.js";
 
@@ -93,4 +95,86 @@ export async function readVerifiedJsonFile(
     return { status: "unreadable", reason: "malformed_json" };
   }
   return { status: "parsed", value };
+}
+
+/**
+ * Where a named Working Copy JSON file stands relative to the Manifest, before
+ * any bytes are read. `ready` carries the verified-file identity; `absent` and
+ * `unavailable` carry a typed reason (prefixed with the caller's `label`) so a
+ * removed-or-missing file stays distinct from a present-but-unusable one.
+ */
+export type FileResolution =
+  | {
+      readonly kind: "ready";
+      readonly file: VerifiedJsonFile;
+      readonly ordinal: number;
+    }
+  | {
+      readonly kind: "absent" | "unavailable";
+      readonly reason: string;
+      readonly ordinal: number | null;
+    };
+
+/**
+ * Resolve a Manifest-declared JSON file within a Source's Working Copy. A file
+ * absent from the Manifest or recorded as absent is `absent`; a file that is
+ * unavailable, uncopied, or missing its size/hash representation is a typed
+ * `unavailable`. Only a `ready` result is safe to pass to `readVerifiedJsonFile`.
+ */
+export function resolveJsonFile(
+  source: CaseSourceRecord,
+  workingCopyPath: string,
+  path: string,
+  label: string,
+): FileResolution {
+  const ordinal = source.entries.findIndex((entry) => entry.path === path);
+  if (ordinal < 0) {
+    return {
+      kind: "absent",
+      reason: `${label}_manifest_entry_missing`,
+      ordinal: null,
+    };
+  }
+  const entry = source.entries[ordinal];
+  if (entry === undefined) {
+    return {
+      kind: "absent",
+      reason: `${label}_manifest_entry_missing`,
+      ordinal: null,
+    };
+  }
+  if (entry.state === "absent") {
+    return { kind: "absent", reason: `${label}_absent`, ordinal };
+  }
+  if (entry.state === "unavailable") {
+    return {
+      kind: "unavailable",
+      reason: `${label}_unavailable:${entry.unavailable_reason ?? "unknown"}`,
+      ordinal,
+    };
+  }
+  if (!entry.copied) {
+    return {
+      kind: "unavailable",
+      reason: `${label}_not_in_working_copy`,
+      ordinal,
+    };
+  }
+  if (entry.size === null || entry.sha256 === null) {
+    return {
+      kind: "unavailable",
+      reason: `${label}_manifest_representation_incomplete`,
+      ordinal,
+    };
+  }
+  return {
+    kind: "ready",
+    ordinal,
+    file: {
+      path: join(workingCopyPath, ...path.split("/")),
+      manifestPath: path,
+      size: entry.size,
+      sha256: entry.sha256,
+    },
+  };
 }
